backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pyjwt==2.9.0
python-multipart==0.0.12
httpx==0.27.2
werkzeug==3.1.3
